# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SAER301 API Documentation",
        "description": "Auto-generated API documentation for SAE301 project",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Production API Server"
        }
    ],
    "paths": {
...
        "/api/team/age-validation": {
            "get": {
                "tags": [
                    "Team Age Validation"
                ],
                "summary": "Show age validation page",
                "description": "Display the team age validation interface with rules and thresholds",
                "operationId": "3928e84999236d41d831270ac8c1c761",
                "responses": {
                    "200": {
                        "description": "Age validation page"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
        "/api/team/validate-ages": {
            "post": {
                "tags": [
                    "Team Age Validation"
                ],
                "summary": "Validate team ages",
                "description": "Validates a team's age composition against the rules: all >= A, and if any < B then at least one >= C",
                "operationId": "ef4e67c19bf5e850a24ce4270b5ff622",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ages": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            14,
                                            15,
                                            20
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "details": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/team/validate-birthdates": {
            "post": {
                "tags": [
                    "Team Age Validation"
                ],
                "summary": "Validate team by birthdates",
                "description": "Validates a team's composition by calculating ages from birthdates",
                "operationId": "d3f7bf9e5890300895dc7dba7e599f0a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "birthdates": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "example": [
                                            "2010-05-15",
                                            "2009-08-22",
                                            "2004-01-10"
                                        ]
                                    },
                                    "reference_date": {
                                        "description": "Date to calculate ages from (defaults to today)",
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Validation result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "valid": {
                                            "type": "boolean"
                                        },
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        },
                                        "details": {
                                            "type": "object"
                                        },
                                        "calculated_ages": {
                                            "type": "array",
                                            "items": {
                                                "type": "integer"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/team/check-participant": {
            "post": {
                "tags": [
                    "Team Age Validation"
                ],
                "summary": "Check participant eligibility",
                "description": "Checks if a single participant meets the minimum age requirement",
                "operationId": "15f1b18edfdddbb8a96ae8c7323c2c8e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "properties": {
                                            "age": {
                                                "type": "integer",
                                                "example": 14
                                            }
                                        },
                                        "type": "object"
                                    },
                                    {
                                        "properties": {
                                            "birthdate": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2010-05-15"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Eligibility result",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "eligible": {
                                            "type": "boolean"
                                        },
                                        "age": {
                                            "type": "integer"
                                        },
                                        "is_minor": {
                                            "type": "boolean"
                                        },
                                        "is_adult": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}